Clean code
 * Remove passing of unexisting parameters
 * Add/Fix PHPDoc
 * Remove dead/useless code
 * Fix arguments/return types
 * Fix static calls of non statics methods

+++ front/stat.global.php
@@ -38,7 +38,7 @@ Session::checkRight("statistic", READ);
 
 if (empty($_GET["date1"]) && empty($_GET["date2"])) {
    $year          = date("Y")-1;
-   $_GET["date1"] = date("Y-m-d", mktime(1, 0, 0, date("m"), date("d"), $year));
+   $_GET["date1"] = date("Y-m-d", mktime(1, 0, 0, (int)date("m"), (int)date("d"), $year));
    $_GET["date2"] = date("Y-m-d");
 }
